razor pages for element added

+++ RedBadgeFinal.Services/BusinessLogic/CharacterServices.cs
@@ -1,9 +1,11 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Http;
 using RedBadgeFinal.Data.Data;
 using RedBadgeFinal.Data.Entities;
 using RedBadgeFinal.Models.CharacterModels;
 using RedBadgeFinal.Services.BusinessLogic.IServices;
-using System.Data.Entity;
+using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
 
 namespace RedBadgeFinal.Services.BusinessLogic
 {
@@ -11,46 +13,86 @@ namespace RedBadgeFinal.Services.BusinessLogic
     {
         private readonly ApplicationDbContext _context;
         private IMapper _mapper;
+        private string _ownerId;
 
-        public CharacterServices(ApplicationDbContext context, IMapper mapper)
+        public CharacterServices(ApplicationDbContext context, IMapper mapper, IHttpContextAccessor httpContextAccessor)
         {
             _context = context;
             _mapper = mapper;
+
+            var userClaims = httpContextAccessor.HttpContext!.User.Identity as ClaimsIdentity;
+            var value = userClaims!.FindFirst("id")?.Value;
+
+            _ownerId = value!;
+
+            if (_ownerId == null)
+                throw new Exception("user is not signed in!!!");
         }
 
         public async Task<bool> CreateCharacter(CharacterCreate model)
         {
             var chara = _mapper.Map<Character>(model);
+            chara.OwnerId = _ownerId;
             await _context.Characters.AddAsync(chara);
             return await _context.SaveChangesAsync() == 1;
         }
 
         public async Task<bool> DeleteCharacter(int id)
         {
-            var chara = await _context.Characters.FindAsync(id);
+            var chara = await _context.Characters
+                .Where(c=>c.OwnerId==_ownerId)
+                .SingleOrDefaultAsync(x=>x.Id==id);
+
             if (chara is null) return false;
+
+            if (chara?.OwnerId != _ownerId) return false;
+
             _context.Characters.Remove(chara);
             return await _context.SaveChangesAsync() == 1;
         }
 
         public async Task<CharacterDetail> GetCharacter(int id)
         {
-            var chara = await _context.Characters.FindAsync(id);
+            var chara = await _context.Characters
+                .Include(c=>c.Weapon)
+                .Include(c=>c.Region)
+                .Include(c=>c.Element)
+                .Where(c => c.OwnerId == _ownerId)
+                .SingleOrDefaultAsync(x => x.Id == id);
+
             if (chara is null) return new CharacterDetail();
+
+            if (chara?.OwnerId != _ownerId) return null!;
 
             return _mapper.Map<CharacterDetail>(chara);
         }
 
         public async Task<List<CharacterListItem>> GetCharacters()
         {
-            var chara = await _context.Characters.ToListAsync();
+            var chara = await _context.Characters
+                .Include(c => c.Weapon)
+                .Include(c => c.Region)
+                .Include(c => c.Element)
+                .Where(c => c.OwnerId == _ownerId)
+                .ToListAsync();
+
+            var noOwnerId = chara.Any(c=>c.OwnerId!= _ownerId);
+            if (noOwnerId)
+                return null;
+            else
             return _mapper.Map<List<CharacterListItem>>(chara);
         }
 
         public async Task<bool> UpdateCharacter(CharacterEdit model)
         {
-            var chara = await _context.Characters.FindAsync(model.Id);
+            var chara = await _context.Characters
+                .Where(c => c.OwnerId == _ownerId)
+                .SingleOrDefaultAsync(x => x.Id == model.Id);
+
             if (chara is null) return false;
+
+            if (chara?.OwnerId != _ownerId) return false;
+
             chara.Name = model.Name;
             chara.Description = model.Description;
             chara.RegionId = model.RegionId;
